refactor(decisions): split ProposalsList into sibling files
ProposalsList.tsx was ~1250 lines. Extract the grid family
(ProposalsGrid, renamed from Proposals), the translation hook, the
filter bar/header, and the skeletons into sibling files, leaving
ProposalsList.tsx as the orchestration core (~446 lines). No behavior
change.
